Split ndb and db fields into separate modules
